Initialisation de l'object RectificationGrid hors du wrapper de triangulation

Closes #1009, #1013, and #1016

See merge request 3d/cars-park/cars!824

+++ cars/orchestrator/cluster/mp_cluster/multiprocessing_cluster.py
@@ -253,10 +253,10 @@ class MultiprocessingCluster(abstract_cluster.AbstractCluster):
         """
         return MpDelayed
 
-    def cleanup(self):
+    def cleanup(self, keep_shared_dir=False):
         """
         Cleanup cluster
-
+        :param keep_shared_dir: do not clean directory of shared objects
         """
 
         # Terminate worker
@@ -269,18 +269,19 @@ class MultiprocessingCluster(abstract_cluster.AbstractCluster):
         self.pool.join()
 
         # clean tmpdir if exists
-        self.wrapper.cleanup()
+        self.wrapper.cleanup(keep_shared_dir=keep_shared_dir)
 
-        if self.tmp_dir is not None:
-            shutil.rmtree(self.tmp_dir)
+        if not keep_shared_dir:
+            if self.tmp_dir is not None:
+                shutil.rmtree(self.tmp_dir)
 
     def scatter(self, data):
         """
         Distribute data through workers
 
-        :param data: task data
+        :param data: data to dump
         """
-        return data
+        return self.wrapper.scatter_obj(data)
 
     def create_task_wrapped(self, func, nout=1):
         """
